--- a/tests/envs/test_augmented_env.py
+++ b/tests/envs/test_augmented_env.py
@@ -1,7 +1,6 @@
 """Tests for AugmentedEnv()."""
 
 from llm4pddl import utils
-from llm4pddl.envs import AUGMENTED_BENCHMARKS
 from llm4pddl.envs.augmented_env import AugmentedEnv
 
 
@@ -12,10 +11,10 @@ def test_augmented_env():
         "num_eval_tasks": 10,
         "train_task_offset": 0
     })
-    for env_name in AUGMENTED_BENCHMARKS:
-        env = AugmentedEnv(env_name)
-        assert env.get_name() == f'augmented-{env_name}'
-        training_tasks = env.get_train_tasks()
-        assert len(training_tasks) == 5
-        eval_tasks = env.get_eval_tasks()
-        assert len(eval_tasks) == 10
+    env_name = "pyperplan-blocks"
+    env = AugmentedEnv(env_name)
+    assert env.get_name() == f'augmented-{env_name}'
+    training_tasks = env.get_train_tasks()
+    assert len(training_tasks) == 5
+    eval_tasks = env.get_eval_tasks()
+    assert len(eval_tasks) == 10