--- a/tests/envs/test_pyperplan_env.py
+++ b/tests/envs/test_pyperplan_env.py
@@ -1,7 +1,6 @@
 """Tests for PyperplanEnv()."""
 
 from llm4pddl import utils
-from llm4pddl.envs import PYPERPLAN_BENCHMARKS
 from llm4pddl.envs.pyperplan_env import PyperplanEnv
 
 
@@ -12,10 +11,10 @@ def test_pyperplan_env():
         "num_eval_tasks": 10,
         "train_task_offset": 0
     })
-    for benchmark_name in PYPERPLAN_BENCHMARKS:
-        env = PyperplanEnv(benchmark_name)
-        assert env.get_name() == f"pyperplan-{benchmark_name}"
-        train_tasks = env.get_train_tasks()
-        assert len(train_tasks) == 5
-        eval_tasks = env.get_eval_tasks()
-        assert len(eval_tasks) == 10
+    benchmark_name = "blocks"
+    env = PyperplanEnv(benchmark_name)
+    assert env.get_name() == f"pyperplan-{benchmark_name}"
+    train_tasks = env.get_train_tasks()
+    assert len(train_tasks) == 5
+    eval_tasks = env.get_eval_tasks()
+    assert len(eval_tasks) == 10