closes #137 

results of `pytest -s tests/ --cov-config=.coveragerc --cov=llm4pddl/ --cov=tests/ --cov-fail-under=100 --cov-report=term-missing:skip-covered --durations=0` now:

```
==================================================================================== slowest durations =====================================================================================
6.08s call     tests/approaches/test_approaches.py::test_create_approach
3.52s call     tests/approaches/test_llm_open_loop_approach.py::test_llm_standard_approach_dynamic_big_example
3.11s call     tests/approaches/test_llm_open_loop_approach.py::test_llm_standard_approach_dynamic_small_example
2.12s call     tests/approaches/test_llm_open_loop_approach.py::test_get_closest_datums
1.94s call     tests/approaches/test_llm_open_loop_approach.py::test_llm_standard_approach[pyperplan-woodworking]
1.89s call     tests/approaches/test_llm_open_loop_approach.py::test_llm_standard_approach_failure_cases[standard]
1.82s call     tests/approaches/test_llm_open_loop_approach.py::test_get_cosine_sim
1.76s call     tests/approaches/test_llm_open_loop_approach.py::test_embed_task
1.69s call     tests/approaches/test_llm_open_loop_approach.py::test_llm_standard_approach[custom-easy_spanner]
1.69s call     tests/approaches/test_llm_planning_approach.py::test_llm_planning_planning_approach
1.67s call     tests/approaches/test_llm_open_loop_approach.py::test_llm_standard_approach_failure_cases[group-by-predicate]
1.59s call     tests/approaches/test_llm_open_loop_approach.py::test_embed_tasks
1.58s call     tests/approaches/test_llm_open_loop_approach.py::test_llm_multi_approach
1.54s call     tests/approaches/test_llm_open_loop_approach.py::test_make_embeddings_mapping
1.48s call     tests/approaches/test_llm_open_loop_approach.py::test_llm_standard_approach[pyperplan-blocks]
1.34s call     tests/test_llm_interface.py::test_openai_llm
0.12s call     tests/test_dataset.py::test_manual_data_generation[pyperplan-gripper]
0.09s call     tests/envs/test_augmented_env.py::test_augmented_env
0.09s call     tests/test_dataset.py::test_manual_data_generation[pyperplan-blocks]
0.07s call     tests/envs/test_custom_env.py::test_custom_env
0.05s call     tests/test_main.py::test_run_pipeline
0.05s call     tests/test_main.py::test_main
0.05s call     tests/test_utils.py::test_validate_plan
0.04s call     tests/approaches/test_pure_planning_approach.py::test_pure_planning_approach
0.04s call     tests/envs/test_envs.py::test_create_env[manual-pyperplan-gripper]
0.03s call     tests/test_utils.py::test_run_planning
0.02s call     tests/test_utils.py::test_pyperplan_problem_to_str
0.01s call     tests/test_main.py::test_run_evaluation

(104 durations < 0.005s hidden.  Use -vv to show these durations.)
=================================================================================== 44 passed in 36.58s ====================================================================================
```